+++ example/Example-Common/src/main/java/ca/lukegrahamlandry/examplemod/model/HudColorConfig.java
@@ -11,7 +11,7 @@ package ca.lukegrahamlandry.examplemod.model;
 
 import ca.lukegrahamlandry.lib.config.Comment;
 
-public class ExampleClientConfig {
+public class HudColorConfig {
     @Comment("colour of the text that displays your kills on the hud")
     public int uiColour = 0xFF0000;
 }